requirements.txt
openai
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
pypdf
sqlalchemy
streamlit
python-dotenv
pandas
matplotlib
scikit-learn
